- Make iterator pack template argument to handle both AttributeIteratorPack and SearchIteratorPack. - Use DocumentWeightOrFilterSearch to implement weightset search when filter and no rank needed.

+++ searchlib/src/vespa/searchlib/queryeval/iterator_pack.h
@@ -31,27 +31,25 @@ public:
     // TODO: use MultiSearch::Children to pass ownership
     SearchIteratorPack(const std::vector<SearchIterator*> &children, MatchDataUP md);
 
-    uint32_t get_docid(uint32_t ref) const {
+    uint32_t get_docid(uint16_t ref) const {
         return _children[ref]->getDocId();
     }
 
-    uint32_t seek(uint32_t ref, uint32_t docid) {
+    uint32_t seek(uint16_t ref, uint32_t docid) {
         _children[ref]->seek(docid);
         return _children[ref]->getDocId();
     }
 
-    int32_t get_weight(uint32_t ref, uint32_t docid) {
+    int32_t get_weight(uint16_t ref, uint32_t docid) {
         _children[ref]->doUnpack(docid);
         return _childMatch[ref]->getWeight();
     }
 
-    void unpack(uint32_t ref, uint32_t docid) {
+    void unpack(uint16_t ref, uint32_t docid) {
         _children[ref]->doUnpack(docid);
     }
 
-    size_t size() const {
-        return _children.size();
-    }
+    uint16_t size() const { return _children.size(); }
     void initRange(uint32_t begin, uint32_t end) {
         for (auto & child: _children) {
             child->initRange(begin, end);
